Replace int with long for resource identifiers across models and repositories

+++ src/Examples/MultiDbContextExample/Repositories/DbContextBRepository.cs
@@ -11,6 +11,6 @@ namespace MultiDbContextExample.Repositories;
 public sealed class DbContextBRepository<TResource>(
     ITargetedFields targetedFields, DbContextResolver<DbContextB> dbContextResolver, IResourceGraph resourceGraph, IResourceFactory resourceFactory,
     IEnumerable<IQueryConstraintProvider> constraintProviders, ILoggerFactory loggerFactory, IResourceDefinitionAccessor resourceDefinitionAccessor)
-    : EntityFrameworkCoreRepository<TResource, int>(targetedFields, dbContextResolver, resourceGraph, resourceFactory, constraintProviders, loggerFactory,
+    : EntityFrameworkCoreRepository<TResource, long>(targetedFields, dbContextResolver, resourceGraph, resourceFactory, constraintProviders, loggerFactory,
         resourceDefinitionAccessor)
-    where TResource : class, IIdentifiable<int>;
+    where TResource : class, IIdentifiable<long>;
